Improve asm language

+++ src/main/java/intcode/assembler/Variable.java
@@ -4,15 +4,15 @@ import java.math.BigInteger;
 
 class Variable implements Parameter, HasAddress {
   private final int len;
-  final int[] values;
+  final BigInteger[] values;
   private int address = -1;
 
   public Variable(int len) {
     this.len = len;
-    this.values = new int[len];
+    this.values = new BigInteger[len];
   }
 
-  public Variable(int[] values) {
+  public Variable(BigInteger[] values) {
     this.len = values.length;
     this.values = values;
   }
